+ Fixed the issue where the double value would not parse into the Long when hitting the back end endpoints

+++ alotta_client/lib/assets/services/api_constants.dart
@@ -32,7 +32,7 @@ class ApiConstants {
     return '$baseUrl$_appUserParam$_update';
   }
 
-  static String getRestaurantsForUserUrl(final double userId) {
+  static String getRestaurantsForUserUrl(final int userId) {
     return '$baseUrl$_userRestaurantsParam?userId=$userId';
   }
 }
